Изменения по начислению капремонта

+++ src/main/java/com/ric/bill/dto/MeterDTO.java
@@ -1,13 +1,11 @@
 package com.ric.bill.dto;
 
 
-import com.ric.bill.model.mt.Meter;
-
 import lombok.Getter;
 import lombok.Setter;
 
 /**
- * DTO для физ счетчика, с его статусом работы 
+ * DTO для физ счетчика, с его статусом работы
  * @author lev
  *
  */
@@ -19,17 +17,17 @@ public class MeterDTO {
 	// лиц.счет
 	private Integer lsk;
 	// статус работы счетчика
-	private Double tp;
+	private Integer tp;
 	// для сортировки (1-неисправен и т.п., 0 -исправен, нет показаний)
-	private Integer ord; 
-	
-	public MeterDTO(Integer lsk, Integer meterId, Double tp, Integer ord) {
+	private Integer ord;
+
+	public MeterDTO(Integer lsk, Integer meterId, Integer tp, Integer ord) {
 		super();
 		this.lsk = lsk;
 		this.meterId = meterId;
 		this.tp = tp;
 		this.ord = ord;
 	}
-	
+
 }
